selfPhoto/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["mtralka <mtralka@umd.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.1"
alembic = "^1.7.7"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = "^0.17.6"
pydantic = {extras = ["email"], version = "^1.9.0"}
psycopg2 = "^2.9.3"
aiofiles = "^0.8.0"
huey = "^2.4.3"
redis = "^4.2.2"
gevent = "^21.12.0"
Pillow = "^9.1.0"
geopy = "^2.2.0"
pyhumps = "^3.5.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.942"
flake8 = "^4.0.1"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
